+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- removed local `kubectl` and `kubernetes` packages in lieu of those from nixpkgs
+
 ## [0.2.0] - 2023-07-07
 
 ### Breaking