+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- default `kubernetes.kubeconfig` to `$HOME/.kube/config`
 - removed local `kubectl` and `kubernetes` packages in lieu of those from nixpkgs
 
 ## [0.2.0] - 2023-07-07