+++ CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.3.0] - 2024-05-05
+
+### Breaking
+
+- removed generated Kubernetes manifest file (`manifest.json`) from default flake package
+
+  See the [documentation](https://kubenix.org/#usage) how to access the generated Kubernetes manifest file
 
 ### Added
 
@@ -14,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - removed local `kubectl` and `kubernetes` packages in lieu of those from nixpkgs
+- pin Bash version of Kubenix CLI script
 
 ## [0.2.0] - 2023-07-07